chore: created the bloc for the stake dialog.
