+++ lib/pages/farm/dialogs/dialogs.dart
@@ -1,3 +1,3 @@
 export 'dual_stake_dialog.dart';
-export 'stake_dialog.dart';
+export 'stake/stake_dialog.dart';
 export 'unstake_dialog.dart';
