+++ lib/pages/farm/widgets/single_logo_farm_title.dart
@@ -1,4 +1,5 @@
 import 'package:ax_dapp/pages/farm/dialogs/dialogs.dart';
+import 'package:ax_dapp/pages/farm/dialogs/stake/bloc/stake_bloc.dart';
 import 'package:ax_dapp/pages/farm/widgets/widgets.dart';
 import 'package:ax_dapp/service/controller/farms/farm_controller.dart';
 import 'package:ax_dapp/service/custom_styles.dart';
@@ -6,6 +7,7 @@ import 'package:ax_dapp/util/toast_extensions.dart';
 import 'package:ax_dapp/wallet/bloc/wallet_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:wallet_repository/wallet_repository.dart';
 
 class SingleLogoFarmTitle extends StatelessWidget {
   const SingleLogoFarmTitle({
@@ -67,9 +69,15 @@ class SingleLogoFarmTitle extends StatelessWidget {
                       if (isWalletConnected) {
                         showDialog<void>(
                           context: context,
-                          builder: (BuildContext builderContext) => StakeDialog(
-                            farm: farm,
-                            layoutWdt: cardWidth,
+                          builder: (BuildContext builderContext) =>
+                              BlocProvider(
+                            create: (context) => StakeBloc(
+                              farmController: farm,
+                            ),
+                            child: StakeDialog(
+                              farm: farm,
+                              layoutWdt: cardWidth,
+                            ),
                           ),
                         );
                       } else {
